requirements.txt
requirements.txt
Flask
psutil
Pillow
Werkzeug
